can delete

+++ mean-course/backend/routes/recaps.js
@@ -67,11 +67,13 @@ checkAuth
     content: req.body.content,
     imagePath: imagePath
   });
-  console.log(recap);
-  Recap.updateOne({ _id: req.params.id }, recap)
+  Recap.updateOne({ _id: req.params.id, creator: req.userData.userId }, recap)
     .then(result => {
-      console.log(result);
-      res.status(200).json({message: 'Update successful!'});
+      if (result.nModified > 0) {
+        res.status(200).json({message: 'Update successful!'});
+      } else {
+        res.status(401).json({message: 'Not authorized!'});
+      }
     });
 });
 
@@ -112,9 +114,12 @@ router.get( "", (req,res,next) => {
 
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Recap.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!"});
+  Recap.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: 'Deletion successful!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   });
 });
